+++ temporian/core/operators/BUILD
@@ -58,15 +58,26 @@ py_library(
 )
 
 py_library(
+    name = "calendar_base",
+    srcs = ["calendar_base.py"],
+    srcs_version = "PY3",
+    deps = [
+        ":base",
+        "//temporian/core/data:dtype",
+        "//temporian/core/data:event",
+        "//temporian/core/data:feature",
+        "//temporian/proto:core_py_proto",
+    ],
+)
+
+py_library(
     name = "calendar_day_of_month",
     srcs = ["calendar_day_of_month.py"],
     srcs_version = "PY3",
     deps = [
-        ":base",
+        ":calendar_base",
         "//temporian/core:operator_lib",
         "//temporian/core/data:event",
-        "//temporian/core/data:feature",
-        "//temporian/proto:core_py_proto",
     ],
 )
 
@@ -75,11 +86,9 @@ py_library(
     srcs = ["calendar_day_of_week.py"],
     srcs_version = "PY3",
     deps = [
-        ":base",
+        ":calendar_base",
         "//temporian/core:operator_lib",
         "//temporian/core/data:event",
-        "//temporian/core/data:feature",
-        "//temporian/proto:core_py_proto",
     ],
 )
